Compilation Bug Solving

+++ ModbusAp.h
@@ -2,7 +2,7 @@
 #define _MODBUSAP_H
 
 #include <stdio.h>
-#include "modbustcp.h"
+#include "ModbusTCP.h"
 #include <string.h>	
 #include <unistd.h>
 #include <sys/time.h>
@@ -10,36 +10,16 @@
 #include <arpa/inet.h>
 
 //-------------------------------------------------------------------
-//                           Código Funções Modbus
+//                      Código Funções Modbus
 //-------------------------------------------------------------------
 #define functionCode_h_regs             0x03
 #define functionCode_w_multRegs         0x10
 
 #define APDU_MAX_LENGTH 253
 
-/**
- * @brief 
- * 
- * @param server_add 
- * @param port 
- * @param st_r 
- * @param n_r 
- * @param val 
- * @return int 
- */
-int Read_h_regs(struct sockaddr_in *server_add , int port, uint32_t st_r , uint16_t n_r , uint16_t *val);
+int Read_h_regs(struct sockaddr_in *server_add , int port, uint16_t st_r , uint16_t n_r , uint16_t *val);
 
-/**
- * @brief 
- * 
- * @param server_add 
- * @param port 
- * @param st_r 
- * @param n_r 
- * @param val 
- * @return int 
- */
-int Write_multiple_regs(struct sockaddr_in *server_add , int port, uint32_t st_r , uint16_t n_r , uint16_t *val);
+int Write_multiple_regs(struct sockaddr_in *server_add , int port, uint16_t st_r , uint16_t n_r , uint16_t *val);
 
 int Get_Request();
